finish hardhat setup
