+++ packages/contracts/evm-contracts/package.json
@@ -37,9 +37,9 @@
   },
   "devDependencies": {
     "@nomicfoundation/hardhat-foundry": "^1.1.2",
-    "@nomicfoundation/hardhat-ignition-viem": "^0.15.5",
+    "@nomicfoundation/hardhat-ignition-viem": "^0.15.6",
     "@nomicfoundation/hardhat-toolbox-viem": "^3.0.0",
-    "@nomicfoundation/ignition-core": "^0.15.5",
+    "@nomicfoundation/ignition-core": "^0.15.6",
     "@openzeppelin/contracts": "^5.0.2",
     "@openzeppelin/contracts-upgradeable": "^5.0.2",
     "hardhat": "^2.22.10",
