+++ packages/contracts/evm-contracts/src/index.ts
@@ -1,2 +1,4 @@
 import './paimaL2.js';
+import './deployment.js';
 export * from './recommendedHardhat.js';
+export * from './deployment.js';
