+++ packages/contracts/evm-contracts/src/recommendedHardhat.ts
@@ -6,10 +6,8 @@ import 'hardhat-dependency-compiler';
 import 'hardhat-interact';
 import 'hardhat-abi-exporter';
 import { TASK_NODE_SERVER_READY } from 'hardhat/builtin-tasks/task-names';
-import { listDeployments } from '@nomicfoundation/ignition-core';
-import * as path from 'path';
-import * as fs from 'fs';
 import * as dotenv from 'dotenv';
+import { copyDeployments } from './deployment';
 
 export function defaultHardhatConfig(config: {
   envPath: string;
@@ -77,6 +75,7 @@ type IgnitionDeployParameters = {
   reset: boolean;
   verify: boolean;
   strategy: string;
+  writeLocalhostDeployment: boolean;
 };
 
 /**
@@ -88,8 +87,6 @@ export function defaultDeployment(
   outDir: string,
   config: IgnitionDeployParameters
 ): void {
-  forceToDisk(rootDir, outDir); // note: need to register this listener first
-
   subtask(TASK_NODE_SERVER_READY, async (_, hre, runSuper) => {
     const result = await runSuper();
 
@@ -100,46 +97,14 @@ export function defaultDeployment(
       },
       config
     );
+    await hre.run(
+      {
+        scope: 'paima',
+        task: 'copy-ignition-deployment',
+      },
+      { rootDir, outDir }
+    );
 
     return result;
   });
-}
-
-/**
- * see https://github.com/NomicFoundation/hardhat-ignition/issues/791
- */
-function forceToDisk(rootDir: string, outDir: string) {
-  scope('ignition').task(
-    'deploy',
-    async (taskArguments: IgnitionDeployParameters, hre, runSuper) => {
-      const network = hre.hardhatArguments.network ?? hre.config.defaultNetwork ?? 'hardhat';
-      // see https://github.com/NomicFoundation/hardhat-ignition/issues/791
-      if (taskArguments.deploymentId == null && network === 'hardhat') {
-        taskArguments.deploymentId = taskArguments.deploymentId ?? 'chain-31337';
-      }
-      const result = await runSuper(taskArguments);
-
-      copyDeployments(rootDir, outDir);
-      return result;
-    }
-  );
-}
-
-export async function copyDeployments(rootDir: string, outDir: string): Promise<void> {
-  const deploymentDir = path.resolve(rootDir, 'src', 'ignition', 'deployments');
-  const deployments = await listDeployments(deploymentDir);
-  for (const deployment of deployments) {
-    const deployedAddressesPath = path.join(deploymentDir, deployment, 'deployed_addresses.json');
-    const json = fs.readFileSync(deployedAddressesPath, 'utf8');
-
-    const deploymentsOutput = path.join(outDir, 'deployments');
-    if (!fs.existsSync(deploymentsOutput)) {
-      fs.mkdirSync(deploymentsOutput);
-    }
-
-    const fixedJson = json.replace(/[\r\n]+$/, '');
-    const outputTs = `export default ${fixedJson} as const;`;
-    // TODO: maybe also generate cjs and mjs files equivalents?
-    fs.writeFileSync(path.join(deploymentsOutput, `${deployment}.ts`), outputTs, { flag: 'w' });
-  }
 }
